Fix draft_orders FK type mismatch: use BIGINT for medication_id, location_id, user FKs
The medications, locations, and users tables use BIGINT primary keys in
Neon, not INTEGER. This caused SQLSTATE 42804 on table creation.

+++ database/create_draft_orders_table.sql
@@ -7,8 +7,8 @@ CREATE TABLE IF NOT EXISTS draft_orders (
     id SERIAL PRIMARY KEY,
 
     -- What medication needs ordering
-    medication_id INTEGER NOT NULL REFERENCES medications(id) ON DELETE CASCADE,
-    location_id INTEGER,
+    medication_id BIGINT NOT NULL REFERENCES medications(id) ON DELETE CASCADE,
+    location_id BIGINT,
 
     -- Snapshot of stock state at generation time
     current_stock_boxes NUMERIC(10,2) NOT NULL DEFAULT 0,
@@ -32,8 +32,8 @@ CREATE TABLE IF NOT EXISTS draft_orders (
         CHECK (status IN ('pending_review', 'approved', 'rejected')),
 
     -- Who did what
-    generated_by INTEGER REFERENCES users(id) ON DELETE SET NULL,
-    approved_by INTEGER REFERENCES users(id) ON DELETE SET NULL,
+    generated_by BIGINT REFERENCES users(id) ON DELETE SET NULL,
+    approved_by BIGINT REFERENCES users(id) ON DELETE SET NULL,
 
     -- Timestamps
     generated_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(),
